♻️ refactor: 既存実装を新ToolManagerに完全置き換え
- 旧ToolManagerAdapter実装を削除
- ToolManagerV2をToolManagerにリネーム
- 後方互換性ラッパーを削除（createDefaultToolManagerのみ維持）
- canvas.tsの型定義を更新

全てのビルドとタイプチェックが成功することを確認済み
